requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
tf-keras
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
async-lru==2.0.5
async-timeout==4.0.3
attrs==25.1.0
babel==2.17.0
backcall==0.2.0
backports.tarfile==1.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
branca==0.8.1
bs4==0.0.2
cachetools==5.5.2
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
cons==0.4.6
contourpy==1.3.0
cryptography==44.0.2
cycler==0.12.1
Cython==3.0.12
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
docstring_parser==0.16
document==1.0
entrypoints==0.4
et_xmlfile==2.0.0
etils==1.12.2
etuples==0.3.9
exceptiongroup==1.2.2
faiss-cpu==1.10.0
fastapi
fastjsonschema==2.21.1
filelock==3.15.4
fire==0.5.0
flatbuffers==24.3.25
folium==0.19.4
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
groq==0.24.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.8
httpimport==1.4.1
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.8
imageio==2.37.0
importlib_metadata==8.7.0
intel-openmp==2021.4.0
ipykernel==6.17.1
ipyleaflet==0.19.2
pdfplumber==0.11.6
ipyparallel==8.8.0
ipython==7.34.0
ipython-genutils==0.2.0
ipython-sql==0.5.0
ipytree==0.2.2
ipywidgets==7.7.1
isoduration==20.11.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
jeepney==0.9.0
Jinja2==3.1.4
jiter==0.9.0
joblib==1.4.2
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-leaflet==0.19.2
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server
jupyter_server_terminals==0.5.3
jupyterlab==4.4.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.14
keyring==25.6.0
kiwisolver==1.4.5
langchain
langchain-community
langchain-core
langchain-experimental
langchain-huggingface
langchain-nvidia-ai-endpoints
langchain-ollama
langchain-openai
langchain-text-splitters
langcodes
langserve
langsmith
language_data==1.3.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.6
lazr.uri==1.0.7
lazy_loader==0.4
libmagic==1.0
logical-unification==0.4.6
loguru==0.7.3
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mkl==2021.4.0
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.3
multipledispatch==1.0.0
mypy_extensions==1.1.0
narwhals==1.27.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.4.1
notebook_shim==0.2.4
numexpr==2.10.2
numpy==1.26.4
oauthlib==3.2.2
ollama==0.4.8
onnxruntime==1.19.0
openai
opencv-python==4.10.0.84
orjson==3.10.18
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
param==2.2.0
parso==0.8.4
pathlib==1.0.1
pdfminer.six==20250327
pdfplumber==0.11.6
pickleshare==0.7.5
pillow==10.4.0
python-magic
platformdirs==4.3.7
playwright==1.48.0
prettytable==3.16.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==5.27.3
psutil==6.0.0
py-cpuinfo==9.0.0
py4j==0.10.9.7
pyarrow==18.1.0
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
pyee==12.0.0
Pygments==2.19.1
pyparsing==3.1.4
pypdfium2==4.30.1
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-utils==3.9.1
pytz==2025.2
pyviz_comms==3.0.4
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
requirements-parser==0.9.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.23.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.14.1
seaborn==0.13.2
SecretStorage==3.3.3
Send2Trash==1.8.3
sentence-transformers==3.4.1
sentencepiece==0.2.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
sqlglot==25.20.2
sqlparse==0.5.3
sse-starlette
starlette
streamlit==1.42.2
streamlit-option-menu==0.4.0
streamlit_folium==0.24.0
sympy==1.13.2
tbb==2021.13.1
tenacity==9.0.0
termcolor==3.1.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
tifffile==2025.3.30
tiktoken==0.9.0
timm==1.0.15
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
toolz==0.12.1
torch==2.3.1
torchvision==0.18.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traittypes==0.2.1
transformers==4.51.3
types-python-dateutil==2.9.0.20241206
types-setuptools==80.0.0.20250429
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.82
ultralytics-thop==2.0.5
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.2
vercel==0.2.1
wadllib==2.0.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==3.6.10
win32_setctime==1.2.0
xyzservices==2025.1.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
